Se agrego modal al salir de la app cerrando sesion y cerrando solo la app, se corrigio el bloqueo de pantalla(scroll) en la carga de los clientes, se corrigio lo de movimiento de apertura de caja, se corrigio carga de ciudades adecuada en cada select, se agregaron mas traducciones

+++ platforms/android/assets/www/js/controllers/app_sidemenu.js
@@ -123,9 +123,33 @@ $ionicModal.fromTemplateUrl('templates/modals/change_language.html', {
 
 
  $scope.logoutHome = function(){
-        console.log(SaveData.removeFolder("config",true));
-        SessionService.destroy();
-        $state.go("login");
+
+    var title = $translate.instant('Logout.Title');
+    var msg = $translate.instant('Logout.Msg');
+    var yes = $translate.instant('Logout.Yes');
+    var no = $translate.instant('Logout.No');
+
+
+            var confirmPopup = $ionicPopup.confirm({
+                   title: ''+title,
+                   template: ''+msg,
+                   cancelText: ''+no,
+                   okText: ''+yes
+                 });
+
+                 confirmPopup.then(function(res) {
+                   if(res) {
+                     console.log(SaveData.removeFolder("config",true));
+                      SessionService.destroy();
+                      $state.go("login");
+                   } else {
+                     console.log('You are not sure');
+                   }
+                 });
+
+
+
+        
     };
 //############################################## FIN CODIGO CERRAR SESION ############################################
